Replaced an incorrect return value with a correct(?) call
This seems to be fixing the main bug/issue I've been having where it attempts to parse punctuation, namely a semicolon, where it shouldn't be. I updated the test suite with more examples, and it passes for all lines of code. Whether or not they are 100% correct AST is up in the air, however.